history_river/dj_backend/requirements.txt
Django>=4.2,<5.0
django-cors-headers>=4.0
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=21.0
python-dotenv>=1.0
Pillow>=10.0
requests>=2.31.0
